Refactor and restructure interface
As per the comments on #516, I've tried to restructure the public API
for the client in a bit of a simpler way. I don't, unfortunately, think
it's possible to simplify this API any further at this point in time.
