+++ rpc/Cargo.toml
@@ -25,9 +25,12 @@ all-features = true
 
 [features]
 default = []
-client = [ "async-trait", "async-tungstenite", "futures", "http", "hyper", "tokio" ]
-http_ws = [ "async-trait", "async-tungstenite", "futures", "http", "hyper", "tokio" ]
-secp256k1 = ["tendermint/secp256k1"]
+client = [ "async-trait", "futures" ]
+secp256k1 = [ "tendermint/secp256k1" ]
+subscription = [ "tokio" ]
+transport_http = [ "http", "hyper", "tokio" ]
+transport_mock = [ "tokio" ]
+transport_websocket = [ "async-tungstenite", "tokio" ]
 
 [dependencies]
 bytes = "0.5"
@@ -39,7 +42,7 @@ tendermint = { version = "0.15.0", path = "../tendermint" }
 thiserror = "1"
 uuid = { version = "0.8", default-features = false }
 
-async-tungstenite = { version="0.5", features = ["tokio-runtime"], optional = true }
+async-tungstenite = { version="0.8", features = ["tokio-runtime"], optional = true }
 async-trait = { version = "0.1", optional = true }
 futures = { version = "0.3", optional = true }
 http = { version = "0.2", optional = true }
