+++ rpc/src/lib.rs
@@ -2,39 +2,51 @@
 //!
 //! ## Client
 //!
-//! Available when specifying the `client` feature flag.
+//! This crate optionally provides access to various levels of client
+//! functionality and transports.
 //!
-//! The RPC client comes in two flavors: a [`MinimalClient`] and a
-//! [`FullClient`]. A `MinimalClient` implementation provides access to all
-//! RPC endpoints with the exception of the [`Event`] subscription ones,
-//! whereas a `FullClient` implementation provides access to all RPC
-//! functionality. The reason for this distinction is because `Event`
-//! subscription usually requires more resources to manage, and may not be
-//! necessary for all applications making use of the Tendermint RPC.
+//! By only specifying the `client` feature flag, all you get access to is the
+//! [`Client`] trait. Additional feature flags for access to different aspects
+//! of the client and different transports are as follows:
 //!
-//! Transport-specific client support is provided by way of additional feature
-//! flags (where right now we only have one transport, but intend on providing
-//! more in future):
+//! | Features | Description |
+//! | -------- | ----------- |
+//! | `client`, `transport_http` | Provides [`HttpClient`], which is a basic RPC client that interacts with remote Tendermint nodes via **JSON-RPC over HTTP**. This client does not provide [`Event`] subscription functionality. See the [Tendermint RPC] for more details. |
+//! | `client`, `subscription`, `transport_websocket` | Provides [`WebSocketSubscriptionClient`], which provides [`Event`] subscription functionality over a WebSocket connection. See the [`/subscribe` endpoint] in the Tendermint RPC for more details. |
+//! | `client`, `transport_mock` | Provides [`MockClient`], which only implements [`Client`] (no subscription functionality) for aiding in testing your application that depends on the Tendermint RPC client. |
+//! | `client`, `subscription`, `transport_mock` | In addition to [`MockClient`], this will give you access to [`MockSubscriptionClient`], which helps you simulate subscriptions to events being generated by a Tendermint node. |
 //!
-//! * `http_ws`: Provides an HTTP interface for request/response interactions
-//!   (see [`HttpClient`]), and a WebSocket-based interface for `Event`
-//!   subscription (see [`HttpWebSocketClient`]).
-//!
-//! [`MinimalClient`]: trait.MinimalClient.html
-//! [`FullClient`]: trait.FullClient.html
-//! [`Event`]: event/struct.Event.html
+//! [`Client`]: trait.Client.html
 //! [`HttpClient`]: struct.HttpClient.html
-//! [`HttpWebSocketClient`]: struct.HttpWebSocketClient.html
+//! [`Event`]: event/struct.Event.html
+//! [`WebSocketSubscriptionClient`]: struct.WebSocketSubscriptionClient.html
+//! [Tendermint RPC]: https://docs.tendermint.com/master/rpc/
+//! [`/subscribe` endpoint]: https://docs.tendermint.com/master/rpc/#/Websocket/subscribe
+//! [`MockClient`]: struct.MockClient.html
+//! [`MockSubscriptionClient`]: struct.MockSubscriptionClient.html
 
 #[cfg(feature = "client")]
 mod client;
+#[cfg(all(feature = "client", feature = "transport_http"))]
+pub use client::HttpClient;
+#[cfg(all(
+    feature = "client",
+    feature = "subscription",
+    feature = "transport_mock"
+))]
+pub use client::MockSubscriptionClient;
+#[cfg(all(
+    feature = "client",
+    feature = "subscription",
+    feature = "transport_websocket"
+))]
+pub use client::WebSocketSubscriptionClient;
 #[cfg(feature = "client")]
-pub use client::{
-    EventRx, EventTx, FullClient, MinimalClient, PendingResultTx, Subscription, SubscriptionId,
-    SubscriptionRouter, DEFAULT_SUBSCRIPTION_BUF_SIZE,
-};
-#[cfg(feature = "http_ws")]
-pub use client::{HttpClient, HttpWebSocketClient};
+pub use client::{Client, ClosableClient};
+#[cfg(all(feature = "client", feature = "transport_mock"))]
+pub use client::{MockClient, MockRequestMatcher, MockRequestMethodMatcher};
+#[cfg(all(feature = "client", feature = "subscription"))]
+pub use client::{Subscription, SubscriptionClient, SubscriptionId};
 
 pub mod endpoint;
 pub mod error;
